Introduce tracer to Open-VSX
To instrument wrap the code to be traced with a try-with-resources block

To run it, use the `logging.properties` file. The output will be in
an /opt/ folder, however, it can be changed.

Once the intermediate trace is written, calling
`jsonify.py output-trace.json cleaned-trace.json` will generate a trace
file capable of being read in Chromium chrome://tracing as well as
Eclipse Trace Compass with the Trace Event plugin installed from the
incubator.

The logger was taken from eclipse trace compass and is EPLv2. It is originally
written by Genevieve Bastien and I.

Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>