Introduce SnapshotHandler
This is a way to take a snapshot of any request that takes over n time.

The SnapshotHandler will consume every event and keep it in memory,
(up to maxEvents). This will then dump the full trace to disk when the
specified threshold is met.

To use the snapshotter, pass the following arguments to OpenVSX when
starting it up:
`-Djava.util.logging.config.file=%gitroot%/logging.properties (where %gitroot% is the directory tracecompass is checked out to)`
The logging.properties should be updated to meet the current configuration of
the Open-VSX server.

Performance impact with ONE user on ONE computer was imperceptible, however,
this needs to be verified prior to deployment in production.

Traces will be placed in the folder specified by `logging.properties`.
The traces are following the trace event standard and can be opened with
chromium or Eclipse Trace Compass.

The logger is based on the Trace Compass logger, which is also EPLv2.

Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>