feat: export smoke test results in prometheus format
In addition to the html report, the smoke test results can now be
exported in prometheus format. The exporter is implemented as a
custom Playwright reporter that writes the results to a file and
optionally sends them to a Prometheus Pushgateway.

The configuration for the Prometheus Pushgateway is read from
environment variables to enable easy integration with CI/CD using
secrets.

This should ensure easy integration of the smoke tests into existing
monitoring solutions by either configuring the Prometheus Pushgateway
in the CI workflow or by scraping the results file from the Github
Actions artifacts.

Additionally, we enable the json reporter for the smoke tests to
provide more easily parseable results for further processing if needed.

All reporters are now configured to save their output to the
`webui//playwright-report` directory to easily distinguish them from
the test result artifacts, such as screenshots and videos, that are
saved in the `webui/test-results` directory.

Contributed on behalf of STMicroelectronics

Signed-off-by: Olaf Lessenich <olessenich@eclipsesource.com>